refactor: Convert dashboard detail layout to server component and extract client logic
