+++ src/app/(after-login)/dashboard/[dashboardid]/_components/DashboardIdSetter.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import { useEffect } from "react";
+import { useDashboardStore } from "@/lib/store/useDashboardStore";
+
+export default function DashboardIdSetter({ id }: { id: string }) {
+  const setDashboardId = useDashboardStore((state) => state.setDashboardId);
+
+  useEffect(() => {
+    setDashboardId(id);
+  }, [id, setDashboardId]);
+
+  return null;
+}
